refactor: 식별자 value class에서 양수 검증 익셉션 제거
