+++ src/main/kotlin/com/server/dpmcore/attendance/application/exception/AttendanceExceptionCode.kt
@@ -1,4 +1,4 @@
-package com.server.dpmcore.attendance.domain.exception
+package com.server.dpmcore.attendance.application.exception
 
 import com.server.dpmcore.common.exception.ExceptionCode
 import org.springframework.http.HttpStatus
