+++ src/main/kotlin/com/server/dpmcore/attendance/application/exception/AttendanceNotFoundException.kt
@@ -1,4 +1,4 @@
-package com.server.dpmcore.attendance.domain.exception
+package com.server.dpmcore.attendance.application.exception
 
 import com.server.dpmcore.common.exception.BusinessException
 import com.server.dpmcore.common.exception.ExceptionCode
