+++ src/main/kotlin/com/server/dpmcore/member/member/domain/model/MemberId.kt
@@ -1,16 +1,8 @@
 package com.server.dpmcore.member.member.domain.model
 
-import com.server.dpmcore.member.member.domain.exception.InvalidMemberIdException
-
 @JvmInline
 value class MemberId(
     val value: Long,
 ) {
-    init {
-        if (value <= 0) {
-            throw InvalidMemberIdException()
-        }
-    }
-
     override fun toString(): String = value.toString()
 }
